# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RejectProductOfferRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectProductOfferRequest",
        "type": "object"
      },
...
      "RfiProductOfferResponse": {
        "properties": {
          "assetTypes": {
            "items": {
              "type": "string"
            },
            "title": "Assettypes",
            "type": "array"
          },
          "classificationLevel": {
            "title": "Classificationlevel",
            "type": "integer"
          },
          "matchReasons": {
            "items": {
              "type": "string"
            },
            "title": "Matchreasons",
            "type": "array"
          },
          "matchScore": {
            "title": "Matchscore",
            "type": "number"
          },
          "offerableToUser": {
            "title": "Offerabletouser",
            "type": "boolean"
          },
          "productId": {
            "format": "uuid",
            "title": "Productid",
            "type": "string"
          },
          "productType": {
            "title": "Producttype",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "rejectionReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejectionreason"
          },
          "releasability": {
            "items": {
              "type": "string"
            },
            "title": "Releasability",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "timePeriodEnd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeperiodend"
          },
          "timePeriodStart": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeperiodstart"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "productId",
          "title",
          "summary",
          "productType",
          "matchScore",
          "matchReasons",
          "classificationLevel",
          "releasability",
          "region",
          "timePeriodStart",
          "timePeriodEnd",
          "assetTypes",
          "offerableToUser",
          "status",
          "rejectionReason"
        ],
        "title": "RfiProductOfferResponse",
        "type": "object"
      },
      "RfiSearchMetricsResponse": {
        "properties": {
          "acceptedProductId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptedproductid"
          },
          "candidateCount": {
            "title": "Candidatecount",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "offeredCount": {
            "title": "Offeredcount",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "rejectedCount": {
            "title": "Rejectedcount",
            "type": "integer"
          },
          "runId": {
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          }
        },
        "required": [
          "runId",
          "query",
          "candidateCount",
          "offeredCount",
          "rejectedCount",
          "acceptedProductId",
          "createdAt"
        ],
        "title": "RfiSearchMetricsResponse",
        "type": "object"
      },
      "RfiSearchResultsResponse": {
        "properties": {
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RfiSearchMetricsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/RfiProductOfferResponse"
            },
            "title": "Offers",
            "type": "array"
          },
          "ticketId": {
            "format": "uuid",
            "title": "Ticketid",
            "type": "string"
          },
          "ticketState": {
            "title": "Ticketstate",
            "type": "string"
          }
        },
        "required": [
          "ticketId",
          "ticketState",
          "offers",
          "metrics"
        ],
        "title": "RfiSearchResultsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Secure intelligence tasking and product orchestration API.",
    "title": "Istari API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rfi-search/{ticket_id}/offers/{product_id}/accept": {
      "post": {
        "operationId": "accept_product_offer_api_v1_rfi_search__ticket_id__offers__product_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RfiSearchResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Product Offer",
        "tags": [
          "rfi-search"
        ]
      }
    },
    "/api/v1/rfi-search/{ticket_id}/offers/{product_id}/reject": {
      "post": {
        "operationId": "reject_product_offer_api_v1_rfi_search__ticket_id__offers__product_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProductOfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RfiSearchResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Product Offer",
        "tags": [
          "rfi-search"
        ]
      }
    },
    "/api/v1/rfi-search/{ticket_id}/results": {
      "get": {
        "operationId": "get_rfi_search_results_api_v1_rfi_search__ticket_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RfiSearchResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rfi Search Results",
        "tags": [
          "rfi-search"
        ]
      }
    },
    "/api/v1/rfi-search/{ticket_id}/run": {
      "post": {
        "operationId": "run_rfi_search_api_v1_rfi_search__ticket_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RfiSearchResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Rfi Search",
        "tags": [
          "rfi-search"
        ]
      }
    },
...
  }
}